Before removing big assets to gain download speed when developping.

+++ Code/Common/Inc/TouchGFX/CMeterTemp.hpp
@@ -36,7 +36,7 @@ class CMeterTemp : public CMeter
 {
 public:
 	CMeterTemp();
-	virtual void display(float value);
+	virtual void display(double value);
 
 private:
 };
